# finance/lib/api-zod/src/generated/types/costAllocation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Financial calculation API for construction consulting
 * OpenAPI spec version: 0.1.0
 */

export interface CostAllocation {
  /** Training & capacity building (%) */
  training: number;
  /** Equipment & software (%) */
  equipment: number;
  /** Office & admin upgrade (%) */
  office: number;
  /** External consultants/contractors (%) */
  consultants: number;
  /** Marketing & business development (%) */
  marketing: number;
  /** Contingency & other (%) */
  other: number;
}