# finance/lib/api-zod/src/generated/types/feeRateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Financial calculation API for construction consulting
 * OpenAPI spec version: 0.1.0
 */

export interface FeeRateInput {
  /** Construction package value (VND, excluding VAT) */
  constructionValue: number;
  /** Equipment cost (VND) */
  equipmentCost?: number;
  /** Construction grade: Cấp đặc biệt | Cấp I | Cấp II | Cấp III | Cấp IV */
  constructionGrade: string;
}