# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Financial calculation API for construction consulting
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  error: string;
}